# use validate2 and pass all tests