--- a/db.js
+++ b/db.js
@@ -1,6 +1,6 @@
 const push = require('push-stream')
 const ssbKeys = require('ssb-keys')
-const validate = require('ssb-validate')
+const validate = require('ssb-validate') // TODO: remove this eventually
 const validate2 = require('ssb-validate2-rsjs')
 const bipf = require('bipf')
 const pull = require('pull-stream')