cc @arj03 @mycognosist

- [x] use validate2
- [x] pass all existing tests
- [x] new base index structure
- [x] addBatch and tests
- [x] addOOOBatch and tests
- [x] use `validate2.ready()`
- [x]  make sure it supports browser, either via -wasm or ssb-validate2 (js)
- [x] OPTIONAL ~~remove ssb-validate1 entirely~~ https://github.com/ssb-ngi-pointer/ssb-validate2-rsjs/issues/31
- [x] OPTIONAL ~~Make -wasm fast~~ https://github.com/ssb-ngi-pointer/ssb-validate2-rsjs/issues/28
- [x] Debouncing batch in EBT
- [x] Test performance on mobile => reduces initial sync duration by 75% :+1:
- [x] `ssb.db.add` to do debouncing and addBatch
- [x] Fix performance of "add 1000 msgs" benchmark
- [x] Consider a license change because ssb-validate2 is AGPL-3.0?
- [x] New version of https://crates.io/crates/ssb-crypto with LGPL-3.0
- [x] New version of https://crates.io/crates/ssb-validate with LGPL-3.0
- [x] New version of https://crates.io/crates/ssb-verify-signatures with LGPL-3.0
  - [x] New version of https://crates.io/crates/ssb-legacy-msg-data with LGPL-3.0
  - [x] New version of https://crates.io/crates/ssb-multiformats with LGPL-3.0
- [x] New version of ssb-validate2-rsjs-node