@arj03 2021-07-05
Just two small comments, overall looks good :), I'll try this out now with the extra 2 functions.