@arj03 2021-07-05
Initial results says that this is 600ms faster than master for 2 feeds in full (3.8s vs 4.4s). That might not sound like a lot, but consider this is including getting the data over the network. This is running against an ancient ssb-server on my vps. Amazing work @mycognosist!