@staltz 2021-07-05
Good to hear about the numbers! @arj03 noticed I used paramap. I just slapped the number 8 there but I think this could be experimented with to find the best value.