@arj03 2021-07-05
I was using this [branch](https://github.com/ssb-ngi-pointer/ssb-db2/tree/validate2-addbatch) to test btw :)