@staltz 2021-07-06
@arj03 This is ready for a code review, but I wouldn't merge it yet (it's a big breaking change, we want to make sure it'll run on browsers and on mobile).