@arj03 2021-07-06
in the browser this is 36 seconds compared to ~5 seconds in node. This is running on battery so the numbers are  a bit higher than the post above.