@staltz 2021-07-07
@arj03 I added two commits to this PR:

* use ssb-validate2 (js) in the browser
* call validate2.ready() on start up