@staltz 2021-07-07
## Mobile benchmarks

I'll use this comment as a wiki where I'll update the benchmark numbers for Manyverse's end-to-end replication benchmark:

- **Before** (each sub item below is an independent run of the benchmarks):
  - 342s
  - 430s
- **After**
  - 407s
  - 233s
  - 485s
  - 499s