@staltz 2021-07-09
I just finished this: 

> - [x] `ssb.db.add` to do debouncing and addBatch