@staltz 2021-07-12
@arj03 The performance of "add 1000 elements" makes much more sense now, but it might still surprise that it's not as fast as before.